Setup holding

+++ Source/Game/Components/PickupComponent.h
@@ -8,17 +8,17 @@ struct PickupCProperties {
 };
 
 struct PickupComponent {
-    int entityId[MAX_ENTITIES];
+    uint16_t entityId[MAX_ENTITIES];
     float range[MAX_ENTITIES];
-    int timer[MAX_ENTITIES];
+    uint16_t timer[MAX_ENTITIES];
 
-    PickupComponent () {
+    PickupComponent() {
         std::fill_n(entityId, MAX_ENTITIES, -1);
         std::fill_n(range, MAX_ENTITIES, 0.0f);
         std::fill_n(timer, MAX_ENTITIES, 0);
     };
-    PickupComponent (const PickupComponent&) = delete;
-    PickupComponent & operator=(const PickupComponent &) = delete;
+    PickupComponent(const PickupComponent&) = delete;
+    PickupComponent& operator=(const PickupComponent&) = delete;
 
     PickupCProperties operator[](int index) {
         return PickupCProperties {
