+++ Source/Game/Entity/Entity.h
@@ -1,6 +1,6 @@
 #pragma once
 #include <bitset>
-const uint32_t MAX_ENTITIES = 128;
+const uint16_t MAX_ENTITIES = 128;
 const uint8_t MAX_COMPONENT_TYPES = 32;
 const uint8_t PLAYER_ENTITY = 0;
